Fix run-test-package with deftest
The last call to `finalize' is actually useless when ones use deftest,
because each test call finalize itself. So what it does is to repeat the
same result that was printed.
Let's change this so we plan/finalize based on the number of tests we're
going to run on the package. This is different from the value of *plan* in
package, since this is based on the number of tests in *tests*.

Signed-off-by: Julien Danjou <julien@danjou.info>